Add files via upload
This project focuses on developing a robust predictive model for the revenue generated by Apprentice Chef, Inc. using a comprehensive dataset. The primary tasks involved in this project include feature engineering, model preparation, variable selection, and model development. The objective is to create a model with high explanatory power while adhering to established guidelines and restrictions. The analysis will be conducted using permitted model types from scikit-learn and statsmodels. The performance of the final model is assessed based on a set of criteria, including train-test gap, response variable usage, model types, code quality, processing time, model output, X-variable usage, and full dataset usage. Violations of these criteria result in penalties deducted from the final model R-Square on the test set. The project ensures that the original dataset remains unaltered, with any feature engineering or transformations performed within the Jupyter Notebook. The final model and candidate models are presented in a well-formatted dynamic string, with proper comments and markdown included for better understanding. The entire code is processed within a 60-second timeframe, showcasing efficient model development and adherence to best practices. The outcome of this project will be a valuable contribution to the revenue prediction capabilities of Apprentice Chef, Inc., and a testament to effective data-driven decision-making.